### 2021-05-09 #3267 @arfon
@bshoshany - thanks for your submission to JOSS. As discussed in the earlier scope discussion in https://github.com/openjournals/joss/issues/918, this is a rather small submission for us so we will send this through our internal scope review.

This may take a couple of weeks to complete so thanks for your patience!
