### 2021-05-14 #3267 @jgostick
Regarding the scope query, this was preemptively discussed [here](https://github.com/openjournals/joss/issues/918)
