### 2021-05-14 #3267 @danielskatz
I'll remove the scope query label, since we already discussed it, and assign it (reluctantly) to @gkthiruvathukal, who volunteered in that earlier discussion
