### 2021-05-16 #3267 @bshoshany
@gkthiruvathukal Here are some potential reviewers who listed C++ as a preferred language and parallel programming as a topic they are comfortable reviewing:
* cedricchevalier19
* davidbeckingsale
* fcdimitr
* jpenuchot
* leonardojaneis
* pitsianis
* RcSepp
* zeroset
