### 2021-05-17 #3291 @whedon
[REVIEW]: A C++17 Thread Pool for High-Performance Scientific Computing
**Submitting author:** @bshoshany (<a href="http://orcid.org/0000-0003-2222-127X">Barak Shoshany</a>)
**Repository:** <a href="https://github.com/bshoshany/thread-pool" target ="_blank">https://github.com/bshoshany/thread-pool</a>
**Version:** v1.5
