### 2021-05-25 16308ec
Fix compilation on MSVC.
Compilation with MSVC fails because of the 'and' keyword, after
replacing with '&&' it works.
