### 2021-06-28 #3291 @gkthiruvathukal
@zeroset You can use the whedon generate command. You should be able to check the boxes. I can check with the EICs if you continue to experience problems with the checklist.
