### 2021-06-29 bshoshany/thread-pool#16 @trevor211
It would be better to provide some unit tests.
As far as I know, there are 2 benifits if you would provide unit tests:

1. Confidence: with all the use cases tested, provider would be more confident with the code.
2. Convenience: it would be convenient for users to refer to the test cases while using the code.
