### 2021-06-29 bshoshany/thread-pool#16 @bshoshany
Thanks for the suggestion! I will add a `.cpp` file to the repository that performs automated tests of each of the package's features in a future release, but probably not before September, since I have a very busy summer.
