### 2021-06-29 #3291 @gkthiruvathukal
@openjournals/joss-eics  Can someone help to fix the problem where @zeroset cannot mark the checkboxes. Do I need to reinvite him as a reviewer?
