### 2021-06-30 #3291 @zeroset
Dear @gkthiruvathukal and @bshoshany unfortunately I have doubts that this submission meets the required substantial scholarly effort. The submitted software contains 238 lines of actual C++ code. The reminder is comments and markdown. Referring to [JOSS Guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) the submission should contain at least 300 lines of code or better 1000 LOC.
