### 2021-06-30 #3291 @bshoshany
Dear @zeroset, the relation between code size and effort with respect to this submission was previously discussed in the [pre-submission enquiry](https://github.com/openjournals/joss/issues/918). There, I explained that making the code as short as possible was an explicit design choice, in order for the thread pool to have the smallest memory footprint possible. That's not a bug, it's a feature that I actually worked quite hard to achieve.

@gkthiruvathukal and @mjsottile both commented in the pre-submission enquiry that code size is not always a measure of effort or utility, and that shorter and more concise implementations reflect a higher degree of effort in order to achieve brevity. @gkthiruvathukal further commented that the effort put into making the very detailed and well put-together documentation should also count in the submission's favor. They therefore suggested that the software should nonetheless be considered to be within the scope of the journal, a suggestion that @danielskatz concurred with. Please see [the full discussion](https://github.com/openjournals/joss/issues/918) for more details.
