### 2021-06-30 #3291 @zeroset
@bshoshany could you provide the code for the performance tests you mention in the readme? If you don't mind i could provide you the performance results on ARM based machine.
