### 2021-06-30 bshoshany/thread-pool#16 @zeroset
I think you should provide some examples and unittests. Especially you should think about automatically executing the tests, maybe in github actions. Unfortunately i am not sure about the available parallelism in github actions. ref [joss-review](https://github.com/openjournals/joss-reviews/issues/3291)
