### 2021-06-30 #3291 @gkthiruvathukal
Thanks, @zeroset. Yes, just to clarify, I was one of the editors who advocated for this submission to be considered. While we have guidelines, submissions are always considered based on intellectual merit and their potential to advance research software.Having worked on lightweight libraries/frameworks for concurrency and threading myself, I saw significant merit in reviewing this submission.

So thanks for proceeding with review!

To our esteemed reviewers, for future reference, you can see the pre-review discussion atop this review issues's thread. This will show the full discussion leading to our decision to review:

https://github.com/openjournals/joss-reviews/issues/3267
