### 2021-07-01 #3291 @bshoshany
> @bshoshany could you provide the code for the performance tests you mention in the readme? If you don't mind i could provide you the performance results on ARM based machine.

Sure, I'd love to be able to add results for additional architectures! The code for the performance tests is in [the repository for my multithreaded matrix class template](https://github.com/bshoshany/multithreaded-matrix). I just updated it with the latest version of that template and the test code. For the tests of the thread pool I took the optional third argument of `measure_operations()` to be `false`, since I did not need to test the performance of the expression templates (which is only relevant for the matrix class itself).

Also, I noticed that you commented in [this issue](https://github.com/bshoshany/thread-pool/issues/16) regarding unit tests for the thread pool class. Were you implying that the unit tests would be needed to pass the review? I was planning to add them in a future release, because I'm a bit busy with other projects right now, but I can try to expedite them if you think they are necessary for the review process.
