### 2021-07-01 #3291 @gkthiruvathukal
@bshoshany From my point of view, I would at least like to see the beginning of unit tests in place, even if there is not 100% coverage. I think unit testing is a necessity for all code, but especially for C/C++ libraries and frameworks. I've been using GoogleTest in my advanced systems programming course, and it's really good (and surprisingly easy to use once you have things set up right). If you can at least get your initial tests in place, I think this would satisfy me and the reviewers.

I have a minimum viable application for Google Test. You might find it useful as a starting point. (It's really basic!) I actually use Google Test (a C++ framework) to test C code, too.

https://github.com/gkthiruvathukal/googletest-mva
