### 2021-07-01 #3291 @zeroset
I think at least some tests are necessary for a software publication. @gkthiruvathukal posted a good start on that one. If you don't want to use googletest you could also have a look at [Catch2](https://github.com/catchorg/Catch2).
For the Performance measurement i would like to see the benchmark script in the repository as well. Without it, the performance results are neither reproducible nor reasonable.
Unfortunately i am also quite busy at the moment, that is why I can only do the review step by step.
