### 2021-07-01 #3291 @bshoshany
Dear @gkthiruvathukal and @zeroset:

> I would at least like to see the beginning of unit tests in place, even if there is not 100% coverage.

That sounds fair. But maybe I can just use existing tests I already wrote, and no third party tools, at least for now?

Like many authors publishing in this journal, I'm a scientist by training, not a professional software developer. This would be the first time I'm developing unit tests, and unfortunately I am not yet familiar with tools such as GoogleTest, Catch2, or GitHub Actions (which @zeroset mentioned in the issue).

As easy as these tools are to use, and as much as I want to advance my software development skills by learning them, it would still take me some time. Developing the tests themselves would also take a considerable amount of time. As I said above, I have other projects to work on this summer (research, course development, grant applications...), some of which have upcoming deadlines, so I won't have time to undertake this task until at least September.

However, there is already a performance test, which I used to generate the results in the documentation, and there is a large number of working code examples in the documentation, which can be used for testing every single function of the thread pool library.

Therefore, would it be okay if, instead of an automated test using third party tools, I will simply combine the existing tests into one big `.cpp` file, which will (automatically) run tests for each feature, and include that file in the repository? I'm afraid that's really the best I can do at the moment, unless you want to postpone the review by a few months 😕

> For the Performance measurement i would like to see the benchmark script in the repository as well. Without it, the performance results are neither reproducible nor reasonable.

It does exist, just in [a different repository](https://github.com/bshoshany/multithreaded-matrix).

There are two projects, the thread pool class and the matrix class, and they are being developed completely independently, so I don't want to merge the repositories. It just so happens that matrix operations are a great way to benchmark multithreading.

All you need to do is have `thread_pool.hpp`, `matrix.hpp`, and `matrix_test.cpp` in the same folder, and you can run the test and get the results, although your results will, of course, be different than mine (I'm looking forward to seeing the results you get on the ARM architecture!)

Since I would prefer to keep the two projects in two separate repositories, would it be sufficient, for the purpose of this review, if I just added the performance test files (`matrix.hpp` and `matrix_test.cpp`) to the "paper" branch of the thread pool repository, but not to the master branch?

Thanks!
