### 2021-07-22 bshoshany/thread-pool#16 @simsong
I am now using `thread-pool` on a soon-to-be-production system. I am creating unit tests and code coverage reported on codecov.io. If you wish, I can send pull requests to this repo with some of them.
