### 2021-07-28 bshoshany/thread-pool#16 @bshoshany
@simsong glad to know that you are using this library in your system! Thanks so much for your suggestion. I would prefer to write my own unit tests (it will be good practice), but I would love to see your unit tests for guidance, if you are willing to provide a link!
