+++ README.md
@@ -6,6 +6,7 @@
 ![GitHub last commit](https://img.shields.io/github/last-commit/bshoshany/thread-pool)
 [![GitHub repo stars](https://img.shields.io/github/stars/bshoshany/thread-pool?style=social)](https://github.com/bshoshany/thread-pool)
 [![Twitter @BarakShoshany](https://img.shields.io/twitter/follow/BarakShoshany?style=social)](https://twitter.com/BarakShoshany)
+[![Open in Visual Studio Code](https://open.vscode.dev/badges/open-in-vscode.svg)](https://open.vscode.dev/bshoshany/thread-pool)
 
 # A C++17 Thread Pool for High-Performance Scientific Computing
 
@@ -16,46 +17,39 @@ Department of Physics, Brock University,\
 Companion paper: [arXiv:2105.00613](https://arxiv.org/abs/2105.00613)\
 DOI: [doi:10.5281/zenodo.4742687](https://doi.org/10.5281/zenodo.4742687)
 
-<!-- TOC depthFrom:2 -->
+* [Abstract](#abstract)
+* [Introduction](#introduction)
+    * [Motivation](#motivation)
+    * [Overview of features](#overview-of-features)
+    * [Compiling and compatibility](#compiling-and-compatibility)
+* [Getting started](#getting-started)
+    * [Including the library](#including-the-library)
+    * [Constructors](#constructors)
+    * [Getting and resetting the number of threads in the pool](#getting-and-resetting-the-number-of-threads-in-the-pool)
+* [Submitting and waiting for tasks](#submitting-and-waiting-for-tasks)
+    * [Submitting tasks to the queue with futures](#submitting-tasks-to-the-queue-with-futures)
+    * [Submitting tasks to the queue without futures](#submitting-tasks-to-the-queue-without-futures)
+    * [Manually waiting for all tasks to complete](#manually-waiting-for-all-tasks-to-complete)
+    * [Parallelizing loops](#parallelizing-loops)
+* [Other features](#other-features)
+    * [Synchronizing printing to an output stream](#synchronizing-printing-to-an-output-stream)
+    * [Setting the worker function's sleep duration](#setting-the-worker-functions-sleep-duration)
+    * [Monitoring the tasks](#monitoring-the-tasks)
+    * [Pausing the workers](#pausing-the-workers)
+* [Exception handling](#exception-handling)
+* [Performance tests](#performance-tests)
+    * [Measuring execution time](#measuring-execution-time)
+    * [AMD Ryzen 9 3900X (24 threads)](#amd-ryzen-9-3900x-24-threads)
+    * [Dual Intel Xeon Gold 6148 (80 threads)](#dual-intel-xeon-gold-6148-80-threads)
+* [Feedback](#feedback)
+* [Copyright and citing](#copyright-and-citing)
 
-- [Abstract](#abstract)
-- [Introduction](#introduction)
-    - [Motivation](#motivation)
-    - [Overview of features](#overview-of-features)
-    - [Compiling and compatibility](#compiling-and-compatibility)
-- [Getting started](#getting-started)
-    - [Including the library](#including-the-library)
-    - [Constructors](#constructors)
-    - [Getting and resetting the number of threads in the pool](#getting-and-resetting-the-number-of-threads-in-the-pool)
-- [Submitting and waiting for tasks](#submitting-and-waiting-for-tasks)
-    - [Submitting tasks to the queue with futures](#submitting-tasks-to-the-queue-with-futures)
-    - [Submitting tasks to the queue without futures](#submitting-tasks-to-the-queue-without-futures)
-    - [Manually waiting for all tasks to complete](#manually-waiting-for-all-tasks-to-complete)
[…]

[…]
