### 2021-07-28 #3291 @bshoshany
Hi @gkthiruvathukal and @zeroset,

I updated the `paper` branch to v1.8 with a few small changes and bug fixes, including an additional line in `paper.md` to indicate that `submit()` now forward exceptions from tasks.

I have not yet received a reply from either of you to my comment from about a month ago. Are you able to proceed with the review without the unit tests, which I will prepare as soon as I have time, but probably not until a couple of months from now?

Thanks,
Barak
