### 2021-07-29 #3291 @gkthiruvathukal
@bshoshany  I would be able to live without unit tests as long as there are some clear examples that can demonstrate the entire API is working properly and the expected outputs being produced.
