### 2021-07-29 #3291 @bshoshany
@gkthiruvathukal sounds fair! I will collect all the examples from `README.md` into one `.cpp` file that you can execute once to test every single aspect of the library. I'll let you know when I've uploaded it. Thanks again!
