### 2021-07-29 #3291 @gkthiruvathukal
@bshoshany Consider the possibility of having separate test programs that show how to use various aspects of the library.  I always like concise examples that serve the dual purpose of performing tests and doing proper documentation. But you can also have an all-in-one driver (perhaps a script) that runs the various examples and ensures they perform the expected mission!
