### 2021-07-29 #3291 @bshoshany
@gkthiruvathukal If it's okay with you, it would be easier for me, for now, to have one source file that runs all the examples instead of separate source files executed one by one via a script - the end result will be the same.

The only caveat is that they're only **examples** and not **automated tests**, so in order to see if they work or not, you will have to look at the output and compare it with what you expect to get. But it would still be a good way to ensure every aspect of the library works properly by just running one command, and will also be great help for me in future updates, to make sure I'm not breaking anything whenever I make a change. The expected outputs are also in the `README.md`, but I will add a separate file with the expected output of the entire test program, for convenience - including the outputs of any new examples I'll add in the process. This will also serve as extra demonstrations for the users. I will automate the tests as soon as I have more time (which will be in a couple of months). Will that be satisfactory for the purpose of the review?
