### 2021-08-12 2d98866
Add doc for vcpkg installation of the library
Since it is available through vcpkg, it could be useful for some users to have the doc about how to install and use the library through this channel.
