+++ thread_pool_test.cpp
@@ -1,0 +1,1112 @@
+// Get rid of annoying MSVC warning.
+#ifdef _MSC_VER
+#define _CRT_SECURE_NO_WARNINGS
+#endif
+
+#include <fstream>
+#include <iomanip>
+#include <random>
+#include <string>
+#include <vector>
+
+#include "thread_pool.hpp"
+
+// Define short names for commonly-used integer types.
+typedef std::int_fast32_t i32;
+typedef std::uint_fast32_t ui32;
+typedef std::int_fast64_t i64;
+typedef std::uint_fast64_t ui64;
+
+// Define two global synced_streams objects: one prints to std::cout and the other to a file.
+synced_stream sync_cout(std::cout);
+std::ofstream log_file;
+synced_stream sync_file(log_file);
+
+// A global thread pool object.
+thread_pool pool;
+
+// A global random_device object used to seed some random number generators.
+std::random_device rd;
+
+// Global variables to measure how many checks succeeded and how many failed.
+ui32 tests_succeeded = 0;
+ui32 tests_failed = 0;
+
+/**
+ * @brief Print any number of items into both std::cout and the log file, syncing both independently.
+ *
+ * @tparam T The types of the items.
+ * @param items The items to print.
+ */
+template <typename... T>
+void dual_print(const T &...items)
+{
+    sync_cout.print(items...);
+    sync_file.print(items...);
+}
+
+/**
+ * @brief Print any number of items into both std::cout and the log file, followed by a newline character, syncing both independently.
+ *
+ * @tparam T The types of the items.
+ * @param items The items to print.
+ */
+template <typename... T>
+void dual_println(const T &...items)
+{
+    dual_print(items..., '\n');
+}
+
+/**
+ * @brief Print a stylized header.
+ *
+ * @param text The text of the header. Will appear between two lines.
+ * @param symbol The symbol to use for the lines. Default is '='.
+ */
+void print_header(const std::string &text, const char &symbol = '=')
+{
+    dual_println();
+    dual_println(std::string(text.length(), symbol));
+    dual_println(text);
+    dual_println(std::string(text.length(), symbol));
+}
+
+/**
+ * @brief Get a string representing the current time.
+ *
+ * @return The string.
+ */
+std::string get_time()
+{
+    const std::time_t t = std::time(nullptr);
+    char time_string[32];
+    std::strftime(time_string, sizeof(time_string), "%Y-%m-%d_%H.%M.%S", std::localtime(&t));
+    return std::string(time_string);
+}
+
+/**
+ * @brief Check if a condition is met, report the result, and count the number of successes and failures.
+ *
+ * @param condition The condition to check.
+ */
+void check(const bool condition)
+{
+    if (condition)
+    {
+        dual_println("-> PASSED!");
+        tests_succeeded++;
+    }
+    else
+    {
+        dual_println("-> FAILED!");
+        tests_failed++;
+    }
+}
+
+/**
+ * @brief Store the ID of the current thread in memory. Waits for a short time to ensure it does not get evaluated by more than one thread.
+ *
+ * @param location A pointer to the location where the thread ID should be stored.
+ */
+void store_ID(std::thread::id *location)
+{
+    *location = std::this_thread::get_id();
+    std::this_thread::sleep_for(std::chrono::milliseconds(10));
+}
+
+/**
+ * @brief Count the number of unique threads in the thread pool to ensure that the correct number of individual threads was created. Pushes a number of tasks equal to four times the thread count into the thread pool, and count the number of unique thread IDs returned by the tasks.
+ */
+ui32 count_unique_threads()
+{
+    std::vector<std::thread::id> thread_IDs(pool.get_thread_count() * 4);
+    for (std::thread::id &id : thread_IDs)
[…]
