+++ thread_pool.hpp
@@ -3,8 +3,8 @@
 /**
  * @file thread_pool.hpp
  * @author Barak Shoshany (baraksh@gmail.com) (http://baraksh.com)
- * @version 1.9
- * @date 2021-07-29
+ * @version 2.0.0
+ * @date 2021-08-14
  * @copyright Copyright (c) 2021 Barak Shoshany. Licensed under the MIT license. If you use this library in published research, please cite it as follows:
  *  - Barak Shoshany, "A C++17 Thread Pool for High-Performance Scientific Computing", doi:10.5281/zenodo.4742687, arXiv:2105.00613 (May 2021)
  *
@@ -12,7 +12,7 @@
  * @details A modern C++17-compatible thread pool implementation, built from scratch with high-performance scientific computing in mind. The thread pool is implemented as a single lightweight and self-contained class, and does not have any dependencies other than the C++17 standard library, thus allowing a great degree of portability. In particular, this implementation does not utilize OpenMP or any other high-level multithreading APIs, and thus gives the programmer precise low-level control over the details of the parallelization, which permits more robust optimizations. The thread pool was extensively tested on both AMD and Intel CPUs with up to 40 cores and 80 threads. Other features include automatic generation of futures and easy parallelization of loops. Two helper classes enable synchronizing printing to an output stream by different threads and measuring execution time for benchmarking purposes. Please visit the GitHub repository at https://github.com/bshoshany/thread-pool for documentation and updates, or to submit feature requests and bug reports.
  */
 
-#define THREAD_POOL_VERSION 1.9
+#define THREAD_POOL_VERSION "v2.0.0 (2021-08-14)"
 
 #include <atomic>      // std::atomic
 #include <chrono>      // std::chrono
@@ -24,8 +24,8 @@
 #include <mutex>       // std::mutex, std::scoped_lock
 #include <queue>       // std::queue
 #include <thread>      // std::this_thread, std::thread
-#include <type_traits> // std::decay_t, std::enable_if_t, std::is_void_v, std::invoke_result_t
-#include <utility>     // std::move, std::swap
+#include <type_traits> // std::common_type_t, std::decay_t, std::enable_if_t, std::is_void_v, std::invoke_result_t
+#include <utility>     // std::move
 
 // ============================================================================================= //
 //                                    Begin class thread_pool                                    //
@@ -36,6 +36,7 @@
 class thread_pool
 {
     typedef std::uint_fast32_t ui32;
+    typedef std::uint_fast64_t ui64;
 
 public:
     // ============================
@@ -72,7 +73,7 @@ public:
      *
      * @return The number of queued tasks.
      */
-    size_t get_tasks_queued() const
+    ui64 get_tasks_queued() const
     {
         const std::scoped_lock lock(queue_mutex);
         return tasks.size();
@@ -109,39 +110,49 @@ public:
     }
 
     /**
-     * @brief Parallelize a loop by splitting it into blocks, submitting each block separately to the thread pool, and waiting for all blocks to finish executing. The loop will be equivalent to: for (T i = first_index; i <= last_index; i++) loop(i);
+     * @brief Parallelize a loop by splitting it into blocks, submitting each block separately to the thread pool, and waiting for all blocks to finish executing. The user supplies a loop function, which will be called once per block and should iterate over the block's range.
      *
-     * @tparam T The type of the loop index. Should be a signed or unsigned integer.
+     * @tparam T1 The type of the first index in the loop. Should be a signed or unsigned integer.
+     * @tparam T2 The type of the index after the last index in the loop. Should be a signed or unsigned integer. If T1 is not the same as T2, a common type will be automatically inferred.
      * @tparam F The type of the function to loop through.
-     * @param first_index The first index in the loop (inclusive).
[…]
