### 2021-08-15 bshoshany/thread-pool#16 @bshoshany
Unit tests are now provided in v2.0.0! Please see `thread_pool_test.cpp`. I am closing this issue, but if you find any problems with the tests, please open a separate issue.
