### 2021-08-15 #3291 @bshoshany
Dear @gkthiruvathukal and @zeroset: I just uploaded a new release, v2.0.0. I had a bit of free time, so I started working on some tests and eventually decided to try to automate them. It was so much fun that I ended up spending 2 full days on it, and I'm very happy with the result!

The file `thread_pool_test.cpp` will perform automated tests of all aspects of the package, and benchmark some multithreaded matrix operations. In addition, the code is thoroughly documented, and is meant to serve as an extensive example of how to properly use the package.

I hope this will be sufficient for you to be able to continue the review! :) Please let me know what you think, and if there's anything else you would like me to do to help the review process.
