### 2021-08-15 #3291 @gkthiruvathukal
@bshoshany I'm on vacation through next week but this seems like a great effort to address concerns raised by @zeroset during review. Please proceed!
