### 2021-08-28 #3291 @gkthiruvathukal
Hi, I am just chiming in briefly here.

@zeroset, your points are valid; however, we already decided to put this through review and we rarely reject. If an author cannot overcome the issues arising from feedback, it is their job to pause or withdraw the submission. The "scholarly effort" may appear borderline, but I'm going to stand by my original assertion (from pre-review) that lightweight frameworks are worthy of consideration. JOSS does not have a novelty requirement when it comes to the software being submitted. We do, however, expect authors to explain how their work will help when it comes to other research software efforts. That is, has this library been used successfully or is there reason to believe it can be used successfully in other projects?

@bshoshany I think it is important to note that @zeroset is recommending reject, but you have a chance to rebut the points and try to make your case. There is also a second reviewer who may or may not share this opinion. I would like to see more discussion happen here and to hear both reviewers follow up to your response. Perhaps you can also give some thought to make it clear who the audience is and at least make mention of the related work and why your approach plays an important role in this discussion. (I, for one, mostly avoid complex frameworks in my own work, especially in lower-power computing environments.)

One area of concern, to me, is going to be the reproducibility issues raised by @zeroset. If this is going to be an emphasis of your work, the results do need to be reported rigorously to support your claims. You may also want to be more clear about what scaling can be expected, realistically, and in what environments. For example, would your approach make sense in non-HPC environments, e.g. edge devices with smaller core counts, etc.?

Let's please keep the discussion going and figure out what we can do to improve the submission. But I will ask that we not reject the submission based on matters discussed in pre-review and whether the work is novel. This is what sets JOSS apart from many other venues. Competing approaches can be accepted with proper rationale and justification.
