### 2021-08-30 #3291 @bshoshany
@gkthiruvathukal Thank you for the clarifications! I'm happy to know that the submission will not be rejected! :)

I believe I have already rebutted all of the points made by @zeroset and made my case in my previous comment.

There are no issues of reproducibility. The fact that matrix addition does not enjoy much of a speedup due to multithreading compared to matrix multiplication is well-known, and is reflected in my own tests as given in `README.md`. On my PC I get 2.3x for addition vs 16.2x for multiplication, on an HPC node I get 13.0x vs. 52.7x, and on @zeroset's system he gets 4.3x vs 76.8x. So the fact that @zeroset also got a large discrepancy between the two operations means that he **did** reproduce the expected results.

As I said above, I'm not entirely sure why matrix addition doesn't enjoy the huge speedup that matrix multiplication enjoys. I've been thinking about it, and my theory is that matrix multiplication, which works by taking the dot product of the rows of the first matrix with the columns of the second matrix, can benefit from first copying the individual row and column to the individual core's L1 cache, and then operating on them independently of the other threads. Copying to the cache takes some time, but negligible compared to the dot product itself.

On the other hand, matrix addition simply adds the elements one after the other, and if the particular range of elements being added is copied to the cache, that doesn't really speed things up since the addition itself is so fast anyway. So the threads are all trying to access different locations in the same memory region at the same time, which significantly impedes performance.

In any case, the lack of speedup in matrix addition is due to limitations of the addition algorithm itself, and (as far as I can tell) does **not** indicate any issues with the thread pool implementation. As you can see, matrix multiplication does achieve almost the maximum speedup possible in both my and @zeroset's tests, which means the thread pool itself is working as expected and there are no issues with its performance.

Regarding the expected scaling, it clearly strongly depends on the algorithm, as the matrix addition vs. multiplication results show. Having more cores makes matrix multiplication proportionally faster, but the same does not apply to matrix addition. With the right algorithm, the user can expect a speedup with any number of cores greater than 1, including low-core-count / non-HPC environments (verified on my laptop with 2 cores and 4 threads).

Here are some improvements I can make to the submission based on your and @zeroset's comments:
1. I can change `paper.md` to put more stress on the fact that the whole point of this thread pool library is that it is simple and lightweight, and compare it with some of the big and complex libraries mentioned by @zeroset.
2. I can add a clarification to `paper.md` that the expected scaling strongly depends on the algorithm, as discussed above, and that the library can work with both low and high core count.
3. I can find another matrix operation (perhaps determinant?) that sees a substantial speedup from multithreading, and add it to the performance test. This will provide additional evidence that the thread pool produces a significant and scalable speedup with the right algorithm.

@gkthiruvathukal and @zeroset, please let me know if you think these improvements will resolve the issues, and if there is anything else I should do to improve the submission. Thanks!
