+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,35 @@
+---
+name: Bug report
+about: Create a report to help improve the library
+title: "[BUG]"
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+
+A clear and concise description of what the bug is.
+
+**Minimal working example**
+
+A short but complete program that can be compiled to reproduce the error. Paste the program between the two code fences. If it's too long or requires multiple files, attach the file(s) instead.
+
+```cpp
+```
+
+**Behavior**
+
+What behavior did you expect to get? What actually happened? If the code failed to compile, please include the full output of the compiler.
+
+**System information**
+
+* CPU model, architecture, # of cores and threads:
+* Operating system:
+* Name and version of C++ compiler:
+* Full command used for compiling, including all compiler flags:
+* Thread pool library version:
+
+(Please note that only the latest version of the thread pool library is supported.)
+
+**Additional information**
