+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,23 @@
+---
+name: Feature request
+about: Want a new feature? Suggest it here.
+title: "[REQ]"
+labels: enhancement
+assignees: bshoshany
+
+---
+
+**Describe the new feature**
+
+A clear and concise description of the feature you want.
+
+**Code example**
+
+An example of code that utilizes the suggested feature. Paste or write it between the two code fences.
+
+```cpp
+```
+
+**Additional information**
+
+Include any additional information here.
