+++ .github/ISSUE_TEMPLATE/failed-test.md
@@ -1,26 +1,0 @@
----
-name: Failed test
-about: The provided automated tests failed on your system? Report it here.
-title: "[TEST]"
-labels: bug
-assignees: bshoshany
-
----
-
-**System information**
-
-* CPU model, architecture, # of cores and threads:
-* Operating system:
-* Name and version of C++ compiler:
-* Full command used for compiling, including all compiler flags:
-* Thread pool library version:
-
-(Please note that only the latest version of the thread pool library is supported.)
-
-**Log file**
-
-Please attach the log file for the automated test to this issue.
-
-**Additional information**
-
-Include any additional information here.
