### 2021-09-12 1ff1244
Delete automated-tests-failed.md
