### 2021-09-12 cb59d8d
Delete pull_request_template.md
