+++ .github/ISSUE_TEMPLATE/pull_request_template.md
@@ -1,20 +1,0 @@
-**Pull request policy (please read)**
-
-> Contributions are always welcome. However, I release my projects in cumulative updates after editing and testing them locally on my system, so my policy is not to accept any pull requests. If you open a pull request, and I decide to incorporate your suggestion into the project, I will first modify your code to comply with the project's coding conventions (formatting, syntax, naming, comments, programming practices, etc.), and perform some tests to ensure that the change doesn't break anything. I will then merge it into the next release of the project, possibly together with some other changes. The new release will also include a note in `CHANGELOG.md` with a link to your pull request, and modifications to the documentation in `README.md` as needed.
-
-**Describe the changes**
-
-What does your pull request fix or add to the library?
-
-**Testing**
-
-Have you tested the new code using the provided automated test program `thread_pool_test.cpp` and/or performed any other tests to ensure that it works correctly? If so, please provide information about the test system(s):
-
-* CPU model, architecture, # of cores and threads:
-* Operating system:
-* Name and version of C++ compiler:
-* Full command used for compiling, including all compiler flags:
-
-**Additional information**
-
-Include any additional information here.
