### 2021-09-15 f012a78
Replaced user-configured sleep time with condition variables. Replaced public "paused" variable with pause()/resume() functions. Adapted tests for changes.
