+++ thread_pool.hpp
@@ -75,8 +75,8 @@ public:
      */
     ui64 get_tasks_queued() const
     {
-        const std::scoped_lock lock(queue_mutex);
-        return tasks.size();
+        const std::scoped_lock lock(tasks_mutex);
+        return get_tasks_queued_unprotected();
     }
 
     /**
@@ -86,7 +86,8 @@ public:
      */
     ui32 get_tasks_running() const
     {
-        return tasks_total - (ui32)get_tasks_queued();
+        const std::scoped_lock lock(tasks_mutex);
+        return get_tasks_running_unprotected();
     }
 
     /**
@@ -96,7 +97,8 @@ public:
      */
     ui32 get_tasks_total() const
     {
-        return tasks_total;
+        const std::scoped_lock lock(tasks_mutex);
+        return get_tasks_total_unprotected();
     }
 
     /**
@@ -144,22 +146,24 @@ public:
             block_size = 1;
             num_blocks = (ui32)total_size > 1 ? (ui32)total_size : 1;
         }
-        std::atomic<ui32> blocks_running = 0;
+        ui32 blocks_running = num_blocks;
+        std::mutex blocks_running_mutex;
+        std::condition_variable loop_done_condition_variable;
         for (ui32 t = 0; t < num_blocks; t++)
         {
             T start = ((T)(t * block_size) + the_first_index);
             T end = (t == num_blocks - 1) ? last_index + 1 : ((T)((t + 1) * block_size) + the_first_index);
-            blocks_running++;
-            push_task([start, end, &loop, &blocks_running]
+            push_task([start, end, &loop, &blocks_running, &blocks_running_mutex, &loop_done_condition_variable]
                       {
                           loop(start, end);
-                          blocks_running--;
+                          std::scoped_lock lock(blocks_running_mutex);
+                          if (--blocks_running == 0) {
+                            loop_done_condition_variable.notify_one();
+                          }
                       });
         }
-        while (blocks_running != 0)
-        {
-            sleep_or_yield();
-        }
+        std::unique_lock<std::mutex> lock(blocks_running_mutex);
+        loop_done_condition_variable.wait(lock, [&]{ return blocks_running == 0; });
     }
 
     /**
@@ -171,11 +175,12 @@ public:
     template <typename F>
     void push_task(const F &task)
     {
-        tasks_total++;
         {
-            const std::scoped_lock lock(queue_mutex);
+            const std::scoped_lock lock(tasks_mutex);
             tasks.push(std::function<void()>(task));
+            tasks_total++;
         }
+        task_available_condition_variable.notify_one();
     }
 
     /**
@@ -201,14 +206,16 @@ public:
      */
     void reset(const ui32 &_thread_count = std::thread::hardware_concurrency())
     {
-        bool was_paused = paused;
-        paused = true;
+        const bool was_paused = paused;
+        pause();
         wait_for_tasks();
         running = false;
         destroy_threads();
         thread_count = _thread_count ? _thread_count : std::thread::hardware_concurrency();
         threads.reset(new std::thread[thread_count]);
-        paused = was_paused;
+        if (!was_paused) {
+          resume();
+        }
         running = true;
         create_threads();
     }
@@ -288,35 +295,33 @@ public:
      */
     void wait_for_tasks()
     {
-        while (true)
-        {
+        std::unique_lock<std::mutex> lock(tasks_mutex);
+        tasks_done_condition_variable.wait(lock, [&]{
             if (!paused)
             {
-                if (tasks_total == 0)
-                    break;
+                return (tasks_total == 0);
             }
             else
             {
-                if (get_tasks_running() == 0)
-                    break;
+                return (get_tasks_running_unprotected() == 0);
             }
[…]
