### 2021-09-16 #3291 @bshoshany
Dear @gkthiruvathukal, @zeroset, and @cedricchevalier19,

To add to my previous post - there is a simpler explanation for why matrix addition isn't seeing a substantial performance boost. Although the thread pool eliminates the substantial overhead of creating and destroying threads, there is still some overhead to the parallelization in the form of splitting the operation into tasks, submitting them to the pool, and executing them by the threads. Since matrix addition is a very simple linear operation that doesn't take much time to execute on its own, this overhead negates much of the benefit of multithreading.

If single-threaded matrix addition takes N ms, then splitting it into T threads means that (ideally) each thread runs for N/T ms. However, if the overhead of parallelization is X ms, then the total runtime will be (N/T + X) ms. If X > N(1 - 1/T) then there is no benefit at all to multithreading. But even if X is smaller, there will still be diminishing returns. This is why we do see some speedup (around 4x in @zeroset's test), but presumably in that case we have X > N(1 - 1/4) which is why we don't see any additional benefit for T>4.

As a test of this hypothesis, we can artificially make matrix addition take longer, for example by repeating the addition operation multiple times for each element:

```cpp
pool.parallelize_loop(
    0, a.rows * a.cols, [&a, &b, &c, &num_add](const ui64 &start, const ui64 &end)
    {
        for (ui64 i = start; i < end; i++)
            for (ui32 n = 0; n < num_add; n++)
                c[i] = a[i] + b[i];
    },
    num_tasks);
```

When I run this test, I find that by increasing `num_add` I can essentially get as much speedup as the CPU can provide me with - in my testing on a 12-core/24-thread CPU I was able to get all the way up to a 19x speedup! The reason this happens is that when the operation takes longer, the overhead becomes negligible relative to the operation itself.

It's been a while since my last post, and I'm still waiting for your reply. Furthermore, it's been 4 months since the start of this review, and I still haven't received any comments from the other reviewer, @cedricchevalier19. I would love to be able to advance this review further, and as I said, I'm more than happy to make improvements to both the paper and the code based on your comments, as detailed in my previous post.

Looking forward to hearing from all of you!
Barak
