### 2021-09-16 #3291 @gkthiruvathukal
@zeroset  I was wondering whether you have any follow up to @bshoshany's follow-up. It seems like Barak is working to address your feedback. Would the improvements he is suggesting (two comments ago) be sufficient to move the needle forward?

Also, we definitely need to hear from @cedricchevalier19. Can you let us know your input on this submission.

@bshoshany  Can you speak more specifically to research projects (your own or others) that are making use of your framework?  I'm assuming the ultimate target for this is not HPC workflows but would like to have a better idea of where this framework makes the most sense.
