### 2021-09-17 #3291 @bshoshany
@gkthiruvathukal thank you so much for the quick reply!

Regarding research projects: The whole reason I started writing this thread pool in the first place was to use in my own research project, which was intended to implement general-relativistic ray-tracing in order to visualize interesting spacetime geometries such as black holes, wormholes, and warp drives. Unfortunately, that project is still in the planning stage, since I never found the time to start working on the ray-tracing code itself :(

The thread pool library has a Zenodo DOI and an arXiv ID, but as far as I know it did not get any citations thus far. Therefore, I am not aware of any specific research projects making use of my library.  However, keep in mind that the library has only been out less than 5 months (the first release was April 24, 2021), and only became popular in the last few months. That's not enough time to publish a paper. Nevertheless, the thread pool is certainly being used by many people - it currently has 226 stars - but I don't know if they are using it for scientific software or for other purposes.

Regarding target: I believe this framework is suitable for use with any kind of multithreaded CPU - from a low-core laptop, through a mid-core desktop, up to a high-core HPC node. Indeed, I tested it and found that it worked well on all three types of systems: my 2-core/4-thread laptop, my 12-core/24-thread desktop, and a 40-core/8-thread Compute Canada node. (Of course, the thread pool is only intended to be used for multithreading on a single computer, not on distributed systems.)

In terms of the target user base, I am mainly aiming at users who may not be too experienced with C++ and/or multithreading, and are looking for a library that is lightweight, simple, easy to use, and quick to learn. This is why it is important to me to have a simple and compact interface, thorough documentation with comprehensive examples, and detailed comments in the code itself. It is also why the library is distributed as a single header file, so that people can start using it immediately, without any complicated setup.
