### 2021-10-04 #3291 @gkthiruvathukal
@bshoshany  Thanks for your follow up. I apologize for the delay on my end as September is the craziest month in my academic year (on average). I'm satisfied with your responses. The number of stars on your project is impressive. It does seem like there are some who are using your library.

We do not require citations of your existing Zenodo archive. One of the goals of JOSS is to improve the likelihood of citations, assuming successful peer review.

@zeroset Can you take a look at @bshoshany's response above and let me know if this moves you either way. I actually see the points you are making. I think the matrix multiplication is not (in general) the most compelling use case for the thread pool in general, which is probably why the results seem a little underwhelming.

@cedricchevalier19 We still have no input from you for this submission. If I don't hear from you soon, I need @bshoshany to help identify a new reviewer to take your place. We'd really benefit from your input at this stage.
