### 2021-10-05 #3291 @cedricchevalier19
I tend to agree with @zeroset for his concern about the "Substantial scholarly effort".

I like the fact that the library is small and self contained, however, I think the minimal "thread pool" goal is somehow missed. By providing advanced features such as `parallelize_loop`, it does more than just thread pooling and thus will be compared against `Kokkos`, `TBB`, `Sycl` or even `OpenMP`. Once we have a `parallel for`, we will expect a `parallel result` ...

Futhermore, C++17 provides parallel semantic through `parallel stl` that is now even implemented for gpus (for example, see http://www.mycpu.org/stdpar-c++/).

To precise my thoughts, I was expecting a library that just performs a smart ordering and caching of threads.

However, I looks like a task-based framework that do not allow to express dependencies. On the thread pooling side, a lot of things can be done such as improving hardware locality. [Starpu](https://team.inria.fr/storm/software/starpu/) provides such thing, but without the nice C++ interface of the reviewed project.

Some more comments about the explanations in the README:

- As mentioned earlier by @zeroset, compiler does not parallelize loops (at least not in C++). They can vectorize but they will not automatically add threads
- Having a 30% improvement thanks to hyper-threading for dense matrix-matrix multiplication is not a great result: it means that the hardware is able to counter the fact that the code is stalling/waiting for data. I do think however, that the whole point of thread pool is to allow efficient tasks change for highly irregular workloads. (By the way, before `Kokkos`, [Trilinos](https://github.com/trilinos/Trilinos) used to have a `ThreadPool` library. It has been removed since !trilinos/Trilinos/pull/3725, I still think that it is a good idea to have one).
- `std::async` is not, strictly speaking, about parallelism. It's a way for waiting one event that will allow the thread to continue. It allows to implement cooperation between threads, a wait being translated by a `yield`.
- There is not clear illustration of thread pooling gains: can we obtain the same performance results without pre-allocating threads ?

To put it in a nutshell, I think this library has a nice API to do basic thread management. However, it does not look particularly well suited for High-Performance Scientific Computing. C++17 parallel stl, `Kokkos`, `Sycl` provide more functionalities and probably more performance portability (dealing with NUMA effects, task dependencies, offloading, ...). A clear statement on how this library compares against state-of-the-art approaches is mandatory for HPC considerations.

However, for small projects, on almost embedded architectures, this library can make sense.

@gkthiruvathukal, sorry again for the delay. I have kind of mixed feelings about this review: the library is interesting, and thanks to @bshoshany efforts easy to use.
However, on a science point of view, there is nothing really new, and I am not sure that the author knows how its work compares to classic alternatives.
Reading one more time https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort, I think this library is a “utility” package that I tend to consider as "minor".
