### 2021-10-05 #3291 @gkthiruvathukal
@cedricchevalier19 Thank you for your detailed review. It is becoming increasingly clear that primary issues in this submission are:

- concerns about substantial scholarly effort
- concerns about the HPC claims

I must admit that I, too, struggled with the second concern (more than the first). I also concur with your and @zeroset's review that the matrix multiplication example does not amount to a compelling use for a thread pool. Even in my own projects, thread pools are best suited to irregular workloads.

The question before us is whether this submission can be accepted. I think it can (with the understanding that some major modifications are truly needed). I must stress that *novelty* is not an absolute requirement for acceptance (or rejection) at JOSS.  I realize that it is a borderline situation in terms of scholarly effort, but we did discuss it in pre-review and determined we could move forward based on the code size and well-put-together documentation.

As I see it, there are a few strengths to this submission, despite the weaknesses identified above.

- delightfully self-contained with minimal dependencies
- nicely documented
- evidence of community interest (260 stars of the project)

One thought I have is that we could give @bshoshany the option of refocusing the submission in a couple of ways, but these might not be agreeable. Nevertheless here is what I think:

- include a proper comparison of related efforts identified by both reviewers (both in the C++ standard & HPC)
- downplay the HPC aspects: I see a significant amount of merit in this work for embedded projects, but this doesn't appear to the the focus.
- focus on a more sparse/irregular workload example instead of matrix multiply, which isn't a great use of thread pools. This might require some significant changes.

My feeling is that all of these would require significant effort by @bshoshany.  I'd also like to know whether @zeroset and @cedricchevalier19 agree with my synthesis. At JOSS, once we put something through peer review, we really try to give authors a chance.
