### 2021-10-06 #3291 @gkthiruvathukal
@cedricchevalier19 Thank you for this follow-up. I also was thinking a sparse-matrix computation is the way to go. I'm pretty sure it would not be difficult to rig up this example. I'd like to hear from @zeroset, too, but hope that my suggestions carve a path forward to get to acceptance.
