### 2021-10-16 #3291 @gkthiruvathukal
@zeroset  Can you please follow up to let us know whether you agree with the changes I proposed above? @cedricchevalier19 agrees with the idea. We would only move toward acceptance if @bshoshany can address all of these points.
