### 2021-10-20 #3291 @arfon
> @openjournals/joss-eics Please note the above decision to withdraw the submission so we can bring this submission to proper closure.

Got it. Thanks. I'll proceed now with the withdrawal on behalf of the author.
